Clean up config juggling routines
